use fixture data for birth registration
